homework1/requirements.txt
Pillow
matplotlib
numpy
torch>=1.6.0
torchvision>=0.7.0
